fix: add logos and working version

+++ models/user.js
@@ -7,7 +7,7 @@ module.exports = function(sequelize, DataTypes) {
       displayName: DataTypes.STRING,
       accessToken: DataTypes.STRING,
       refreshToken: DataTypes.STRING,
-      identification: DataTypes.BIGINT(255),
+      identification: DataTypes.INTEGER,
     },
     {
       classMethods: {
